Relocate files and adjust imports
